More massive changes!
Most importantly of these changes is that the http plugin will now find the cache
plugin if it's loaded, and will cache URL's to disk. There is a problem with it, at
the moment, that I intend to fix by this weekend at the latest: right now it is highly
memory intensive. Here's the low-down:

* When a URL is retrieved, it's stored entirely in RAM and then appended to a BMessage. Bad.
* After appending it to the BMessage, the message is then forwarded to the cache plugin
  which currently makes a copy of the message, then processes it. Worse.
* After the message has been processed, an entry_ref to the cache file is sent to the
  http plugin which then actually stores the data. So so.
* A test download of a 21MB file used 63MB RAM total for the transaction. Dear God!

The fix is simple, and I'll do it today or tomorrow. (Yeah!)

At this point there still isn't even the slightest bit of rendering going on, although
I'm still working on the structures to do just that. I hope to have them in place soon.

Cached files typically can be viewed with the appropriate sort of viewer/editor:
	MPEGs can be played with MediaPlayer (provided you drag the file into an open window)
	Images viewed with ShowImage (same thing)
	etc.

The http protocol is still far from officially 1.1 compliant, but it's generally usable.
I'm planning on adding support for chunked server responses next along with support for
streaming data, no-cache response headers, and some other minor things within the next
week or so.

-= Z3R0 One =-


git-svn-id: svn://svn.code.sf.net/p/themis/code@36 ecaf47ba-bb13-0410-a3b6-899b003a5108