Updated the project files a bit, plus a few more fixes here and there:
The project files will now try to create their respective binaries at Themis/...
so make sure the appropriate folders exist. (Alternately, build the project with
the makefile first or exclusively.) Building the project through its various
project files is currently the easiest way of getting the debug information
included in them.

Fixed a bug in httplayer.cpp that triggered a crash when the web server sent
an incomplete header block to Themis. What would happen is that the server would
send part of the headers (with the last bytes being "\r\n") and the ProcessHeaders()
function therefore couldn't find the end of the header block, resulting in eoh being
set to NULL. It would later then try to use this value... Yuck... The fix is simply
to keep receiving data until the header terminator "\r\n\r\n" is found, then process
the headers.

Added a blank line at the end of optionshandler.cpp to prevent a simple but annoying
warning from popping up.

Added #include <stdlib.h> to httpheader.h  to avoid an error message that wasn't
occurring on my system.


git-svn-id: svn://svn.code.sf.net/p/themis/code@57 ecaf47ba-bb13-0410-a3b6-899b003a5108