Added a few more targets, including cleaning ones. See "make help" for details. Fixed dependency handling a bit; sometimes the dependency files were being regenerated every time through, causing the object files to be rebuilt --- which defeats the purpose of dependency handling.
Note that dependencies are still slightly broken: sometimes modifying a
header file will not trigger the dependency file to get updated, meaning
that the appropriate object file may not get updated which in turn means
that you'll get link errors. Thus the new cleaning targets.

In addition, there are javascript and cookie targets, though the code for
these has not been submitted. I've commented out the parts of the makefile
that would automatically try to build these parts, but you can still attempt
to "make them manually". Of course, you don't have the code for them so that
would be rather pointless. :-)

ZO


git-svn-id: svn://svn.code.sf.net/p/themis/code@69 ecaf47ba-bb13-0410-a3b6-899b003a5108