Forgot cardinal rule of modifying a makefile:
Always try your changes after doing a make clean, but before committing
them to the repository.

Duh... Simple change to make things continue to work right for the project
as is. This undoes the header sensitivity for the framework application
though the add-ons and libraries still have it.

The reason is that the change I made "make" look for a header file of almost
the same name as the source file to update dependencies. If no such header
file exists, the particular dependency file isn't built, thus the object file
isn't built, thus the target isn't built, thus you become one unhappy camper.

This did not affect any target other than "framework" so the fix was simply
to remove the need to have a .h for the dependency files for the framework
target.

ZO


git-svn-id: svn://svn.code.sf.net/p/themis/code@116 ecaf47ba-bb13-0410-a3b6-899b003a5108