Many, many changes.
Complete rewrite of the plug-in manager, including now operable node watching
plug-in loading/unloading. This new rewrite supports the broadcast mechanism
discussed previously.

In addition, I have moved the plugman.cpp file from the framework directory
to the common directory, where it truly belongs.

Please be aware that this is incomplete but functional work!

There is still much work to be done in the add-ons to make use of the
broadcast mechanism, and there will eventually be more subclasses of the
PlugClass base class.

This is a work in progress. Don't mind the dust, but feel free to grab a
hammer and fix something.

Raymond


git-svn-id: svn://svn.code.sf.net/p/themis/code@120 ecaf47ba-bb13-0410-a3b6-899b003a5108