A lot of changes that aren't terribly well described at the moment...
The biggest news is that Themis will now display images in a separate window
if the image isn't already cached (a little more work needs to be done) and
the testplug has been compiled.

The http layer has also been adjusted so that when it receives data, it will
broadcast that it received data. It includes the number of bytes received, and,
if available, the content length. This now provides an easy way of discovering
whether the entire file has been downloaded, while providing a mechanism for
progressive updating. (Like progressive jpegs, movies, etc.)

other changes to be described later... I have to run!
Raymond


git-svn-id: svn://svn.code.sf.net/p/themis/code@127 ecaf47ba-bb13-0410-a3b6-899b003a5108