Committing in .
Modified Files:
	FakeSite.cpp FakeSite.h ThemisStatusView.cpp
	ThemisStatusView.h ThemisTVS.h ThemisTabView.cpp
 	ThemisUrlView.cpp ThemisUrlView.h win.cpp win.h

I found some time and put in some progressbars. All progress in loading
is just faked. And I have not tested everything that much. I suppose
there are stil bugs and things which can be done better. But for the
moment... If the parsers are working fine and we are beginning loading
images while parsing, we can go for sending real document/image loading
messages. Good night.


git-svn-id: svn://svn.code.sf.net/p/themis/code@203 ecaf47ba-bb13-0410-a3b6-899b003a5108