Wow!! Massive changes!!
Well, most importantly amongst these changes is the new networking system. It
is disabled by default (you need to modify line 251 of the makefile so that -DNEWNET
is on the left of the # symbol) but it should increase Themis' performance considerably,
and make it a lot more stable. It needs to be polished, and have a lot of printf's
removed, but it should compile as is on both BONE (with R5 and Dano/Exp) and net_server.
In theory, it should also compile on Zeta, though I don't know for sure, though R5+BONE
binaries did run on Zeta.

Also, I've tweaked the http layer a bit, and have added a new implementation of base-64
encoding and decoding in class form. This functionality is not yet in use, but it's
present nonetheless. Actually, I should probably move it to the common directory, but
for now I want it in the repository. It was implemented based on a specification whose
number I forget for another of my projects but should work quite well.

The test plug-in has been stripped back down to the basics; it will now simply
print out the content of any completed HTTP request from cache. This is useful
if you're not sure that Themis is getting all of a site's data. The simplified code
is also much easier to read.

There are still many problems in the old code as well as the new. I'm working towards
resolving as many as possible.

Raymond


git-svn-id: svn://svn.code.sf.net/p/themis/code@233 ecaf47ba-bb13-0410-a3b6-899b003a5108