Applied memleakfixes to the code were possible. All the credits go to Olivier! I was just the one who produced them, and synced them. Thanks Olivier! ( AboutWindow changes have not been applied yet. )
And some other little memleak fixes I found on my own, yeah :)

Prefswin Newpage selection fix for post R5 systems.

Modified makefile that Themis links against libzeta.so in Zeta.
This is necessary, as original find_directory() was not thread safe,
and it had been rewritten as find_directory_r() to which is redirected
when calling find_directory() on Zeta.

Michael


git-svn-id: svn://svn.code.sf.net/p/themis/code@240 ecaf47ba-bb13-0410-a3b6-899b003a5108