Reverted changes made to the makefile involving the renderer. I don't feel that we are quite on the path to having the renderer yet, and I don't want it a part of the official makefile or built by default until we have a functional one yet. That's why I sent the modified makefile by e-mail rather than commiting the changes to CVS.
My apologies if this angers anyone, but there are a fair number of anonymous CVS users
that might get overly excited if they see the renderer being built when the rest of Themis
is built, but don't actually get to see anything rendered. If you wish to continue to use
the modified makefile, be sure to commit from within the modules/Renderer directory or whatever
directory you have done the majority of your changes in. This will prevent the modified makefile
from getting into the repository again.

Raymond


git-svn-id: svn://svn.code.sf.net/p/themis/code@255 ecaf47ba-bb13-0410-a3b6-899b003a5108