A few relatively minor changes that have major impacts on some areas of functionality.
Added a Debug function to the message system to simplify the efforts of getting
debugging messages to the MessageViewer add-on. Also added CountMembers() and
GetMember() to the message system to be able to quickly and easily access message
system members. In addition added MsgSysObjectName() so that you may get the name
of any given message system member. This means we need to start passing the message
system name parameter in all constructors of classes derived directly or indirectly
from MessageSystem (including classes derived from PlugClass).

I updated that in a few locations.

Also I touched but made no change to TRendererEngine, so it's getting included in
the commit because of cvs's functionality. I made a minor change to imageman.cpp as well,
that has no functionality impact, so it's not worth mentioning yet.

Raymond


git-svn-id: svn://svn.code.sf.net/p/themis/code@366 ecaf47ba-bb13-0410-a3b6-899b003a5108