SiteEntrys state gets set now to 100, if all its child UrlEntries are loaded. Additionally the Statustext ist set to "Done." now, as it used to be in earlier days ;)
Fixed FileProtocol to return site/url_id combination instead of view_id ;)

Additionally changed to use the "url" which is comming back from the proto,
instead of relying on the one stored in the Site/UrlEntry. Fileproto
acrually returned the filesystem URL without proto, as the file is
a disk cache file. So now parsing works. Renderer does Render, but still
goes the view_id way. That's why nothing is displayed.

Michael


git-svn-id: svn://svn.code.sf.net/p/themis/code@395 ecaf47ba-bb13-0410-a3b6-899b003a5108